Tree: Limit each file to one component/proptypes
This helps documentation tools that parse AST and detect multiple proptypes and functional component renders.
